refactor(#222): hoist generate_batch_sql and rows_to_maps into smugglr-core
The two pure functions were byte-identical duplicates in the http-sql
plugin adapter and the wasm adapter_common module. Hoisted them into a
new smugglr-core::batch_sql module (always compiled, no native deps,
same pattern as rowhash.rs) and updated both adapters to call the
shared implementation. A snapshot test pins the exact emitted SQL
string so the hoist cannot silently change adapter behavior.
