+++ crates/smugglr-core/src/lib.rs
@@ -9,6 +9,7 @@
 //! reached through `PluginDataSource`. Without `native`, only the diff/sync
 //! engine and trait definitions are available.
 
+pub mod batch_sql;
 pub mod config;
 pub mod datasource;
 pub mod diff;
